added migration

+++ HotelBookingSystem/src/HotelBookingSystem.Domain/Entities/UserRole.cs
@@ -8,9 +8,9 @@ namespace HotelBookingSystem.Domain.Entities;
 
 public class UserRole
 {
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; }
 
-    public int RoleId { get; set; }
+    public long RoleId { get; set; }
     public Role Role { get; set; }
 }
